requirements.txt
# Core Dependencies
pillow>=9.0.0
img2pdf>=0.4.0

# Testing Framework
pytest>=7.0.0
pytest-cov>=3.0.0

# GUI Frameworks
PyQt5>=5.15.0

# Enhanced Features
numpy>=1.21.0
colorama>=0.4.4

# Optional dependencies for enhanced features:
# tkinterdnd2>=0.3.0        # For drag and drop support
# opencv-python>=4.5.0     # For advanced image processing
# matplotlib>=3.5.0        # For image analysis and previews
# reportlab>=3.6.0         # For advanced PDF features
